refactor: decompose FSDP/Megatron into Mixin architecture
Split monolithic engine classes into modular Mixins (dist/state/compute)
following TrainEngine API hierarchy. Glue classes now only contain __init__,
combining Mixins via multiple inheritance.

- Add fsdp/{dist,state,compute,protocol}.py
- Add megatron/{dist,state,compute,protocol}.py
- Simplify fsdp_engine.py and megatron_engine.py to glue classes
