+++ areal/api/engine_api.py
@@ -5,473 +5,33 @@ from collections.abc import Callable
 from concurrent.futures import Future
 from typing import TYPE_CHECKING, Any
 
-import torch
-import torch.distributed as dist
 from torchdata.stateful_dataloader import StatefulDataLoader
 
-from areal.api.alloc_mode import ParallelStrategy
 from areal.api.io_struct import (
     LocalInfServerInfo,
     ModelRequest,
     ModelResponse,
     ParamSpec,
-    SaveLoadMeta,
     WeightUpdateMeta,
+)
+from areal.api.train_engine import (
+    TrainEngineComputeMixin,
+    TrainEngineDistMixin,
+    TrainEngineStateMixin,
 )
 
 if TYPE_CHECKING:
     from areal.api.workflow_api import RolloutWorkflow
     from areal.core.workflow_executor import WorkflowExecutor
-    from areal.utils.data import MicroBatchList
 
 
-class TrainEngine(abc.ABC):
-    @abc.abstractmethod
-    def create_process_group(self, parallel_strategy: ParallelStrategy | None = None):
-        """Initialize PyTorch distributed communication groups.
-
-        Parameters
-        ----------
-        parallel_strategy : ParallelStrategy, optional
-            The parallel strategy configuration for distributed training, by default None
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def initialize(self, *args, **kwargs):
-        """Initialize environments for distributed training and load models.
-
-        This method should be called after `create_process_group`.
-
-        Parameters
-        ----------
-        *args
-            Variable length argument list
-        **kwargs
-            Arbitrary keyword arguments
-        """
-        raise NotImplementedError()
-
-    @property
-    @abc.abstractmethod
-    def data_parallel_group(self) -> dist.ProcessGroup:
-        """Get the data parallel communication group of this engine.
-
-        Returns
-        -------
-        dist.ProcessGroup
-            The data parallel communication group
-        """
-        raise NotImplementedError()
-
-    @property
-    @abc.abstractmethod
-    def data_parallel_rank(self) -> int:
-        """Get the rank of the current process in the data parallel group.
-
-        Returns
-        -------
-        int
-            The rank of the current process in the data parallel group
-        """
-        raise NotImplementedError()
-
-    @property
-    @abc.abstractmethod
-    def data_parallel_world_size(self) -> int:
-        """Get the world size of the data parallel group.
-
-        Returns
-        -------
-        int
-            The world size of the data parallel group
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def current_data_parallel_head(self) -> int:
-        """Get the current data parallel head rank.
-
-        Returns
-        -------
-        int
-            The rank of the current data parallel head
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def is_data_parallel_head(self) -> bool:
-        """Check if the current rank is the data parallel head of the current engine.
-
-        Returns
-        -------
-        bool
-            True if the current rank is the data parallel head, False otherwise
-        """
-        raise NotImplementedError()
-
-    @property
-    @abc.abstractmethod
-    def context_and_model_parallel_group(self) -> dist.ProcessGroup:
-        """Get the context and model parallel communication group of this engine.
-
-        Returns
-        -------
-        dist.ProcessGroup
-            The context and model parallel communication group
-        """
-        raise NotImplementedError()
-
-    @property
-    @abc.abstractmethod
-    def cpu_group(self) -> dist.ProcessGroup:
-        """Get the CPU communication group of this engine.
-
-        Returns
-        -------
-        dist.ProcessGroup
-            The CPU communication group
-        """
-        raise NotImplementedError()
-
-    def destroy(self):
-        """Destroy the engine and release GPU memory of models."""
-
-    @abc.abstractmethod
-    def train(self, mode: bool = True):
-        """Set the engine to training mode.
-
-        Parameters
-        ----------
-        mode : bool, optional
-            Whether to set the engine to training mode, by default True
-        """
-        raise NotImplementedError()
-
-    def eval(self):
-        """Set the engine to evaluation mode.
-
-        This is a convenience method that calls `self.train(False)`.
-        """
-        return self.train(False)
-
-    @abc.abstractmethod
-    def update_weights(self, meta: WeightUpdateMeta):
-        """Update weights to the inference engine in a blocking manner.
-
-        Parameters
-        ----------
-        meta : WeightUpdateMeta
-            Metadata containing information about the weight update
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def connect_engine(self, engine: InferenceEngine, meta: WeightUpdateMeta):
-        """Connect to an inference engine for online training.
-
-        Parameters
-        ----------
-        engine : InferenceEngine
-            The inference engine to connect to
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def rollout_batch(
-        self,
-        data: list[dict[str, Any]],
-        granularity: int = 1,
-        workflow: RolloutWorkflow | type[RolloutWorkflow] | str | None = None,
-        workflow_kwargs: dict[str, Any] | None = None,
-    ) -> dict[str, Any]:
-        """Submit a batch of requests and wait for results.
-
-        This method does not support asynchronous rollout and should be used for offline
-        data collection or debugging, not in production experiments.
-        Should note that this is a simple rollout engine method forwarding with
-        distributed data management.
-
-        Parameters
-        ----------
-        data : list[dict[str, Any]]
-            A list of input data dictionaries.
-        granularity : int, optional
-            The granularity of the rollout, by default 1.
-        workflow : RolloutWorkflow | type[RolloutWorkflow] | str | None, optional
-            The workflow to use for rollout generation, by default None.
-        workflow_kwargs : dict[str, Any] | None, optional
-            Keyword arguments to pass to the workflow constructor, by default None.
-
-        Returns
-        -------
-        dict[str, Any]
-            The rollout results.
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def prepare_batch(
-        self,
-        dataloader: StatefulDataLoader,
-        granularity: int = 1,
-        workflow: RolloutWorkflow | type[RolloutWorkflow] | str | None = None,
-        workflow_kwargs: dict[str, Any] | None = None,
-        should_accept_fn: Callable[[dict[str, Any]], bool] | str | None = None,
-    ) -> dict[str, Any]:
-        """Prepare a batch of data for training from a dataloader.
-
-        Parameters
-        ----------
-        dataloader : StatefulDataLoader
-            The dataloader to fetch data from.
-        granularity : int, optional
-            The granularity of the rollout, by default 1.
-        workflow : RolloutWorkflow | type[RolloutWorkflow] | str | None, optional
-            The workflow to use for rollout generation, by default None.
-        workflow_kwargs : dict[str, Any] | None, optional
-            Keyword arguments to pass to the workflow constructor, by default None.
-        should_accept_fn : Callable[[dict[str, Any]], bool] | str | None, optional
-            A function to filter trajectories, by default None.
-
-        Returns
-        -------
-        dict[str, Any]
-            The prepared batch data.
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def set_version(self, version: int):
-        """Set the current weight version in the training engine.
-
-        Parameters
-        ----------
-        version : int
-            The weight version number to set
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def get_version(self) -> int:
-        """Get the current weight version in the training engine.
-
-        Returns
-        -------
-        int
-            The current weight version number
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def save(self, meta: SaveLoadMeta):
-        """Save model weights and optimizer states for later use.
-
-        Parameters
-        ----------
-        meta : SaveLoadMeta
-            Metadata containing information about where and how to save
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def load(self, meta: SaveLoadMeta):
-        """Load model weights and optimizer states from a file.
-
-        Parameters
-        ----------
-        meta : SaveLoadMeta
-            Metadata containing information about where and how to load
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def optimizer_zero_grad(self):
-        """Zero out all gradients in the optimizer."""
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def optimizer_step(self):
-        """Perform a single optimization step.
-
-        Returns
-        -------
-        dict[str, float]
-            Training statistics containing ``update_successful``, ``grad_norm``, and ``lr``.
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def lr_scheduler_step(self):
-        """Advance the learning rate scheduler by one step."""
-        raise NotImplementedError()
-
-    def step_lr_scheduler(self):
-        """This is an alias for `lr_scheduler_step()`."""
-        return self.lr_scheduler_step()
-
-    @abc.abstractmethod
-    def forward_backward_batch(
-        self,
-        mb_list: MicroBatchList,
-        process_output_fn: Callable[
-            [torch.Tensor, dict[str, Any]], torch.Tensor | None
-        ],
-        forward_only: bool = False,
-    ) -> None:
-        """Process micro-batches through forward and optionally backward pass.
-
-        Parameters
-        ----------
-        mb_list : MicroBatchList
-            The micro-batch list, which is iterable and yields MicroBatchItem tuples.
-        process_output_fn : Callable[[torch.Tensor, dict[str, Any]], torch.Tensor | None]
-            A function that processes the model output (logits) and returns the loss tensor.
-            If the returned loss is not None, backward() will be called on it.
-            Results can be collected via closure if needed.
-            Signature: ``(logits: Tensor, inputs: dict) -> loss | None``
-        forward_only : bool, optional
-            If True, skip backward pass. Default is False.
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def train_batch(
-        self,
-        input_: dict[str, Any],
-        loss_fn: Callable[..., torch.Tensor],
-        loss_weight_fn: Callable[[dict[str, Any]], torch.Tensor],
-    ) -> dict[str, float]:
-        """Update the model with a batch of data and a loss function.
-
-        Note
-        ----
-        The loss_fn should process packed 1D inputs, instead of 2D inputs.
-
-        Parameters
-        ----------
-        input_ : dict[str, Any]
-            The input data for model forward pass and the loss function.
-            Redundant entries are allowed.
-        loss_fn : Callable[..., torch.Tensor]
-            The loss function. For actor (is_critic=False), it receives
-            (logprobs, entropy, input_data). For critic (is_critic=True),
-            it receives (values, input_data). Returns a scalar normalized loss.
-        loss_weight_fn : Callable[[dict[str, Any]], torch.Tensor]
-            A function used to calculate the weight of each micro-batch. Since
-            loss_fn normalizes the loss for a micro-batch, we need a corresponding
-            weight for each micro-batch to normalize the loss globally. The weight
-            is usually the number of response tokens in the batch.
-
-        Returns
-        -------
-        dict[str, float]
-            Scalar statistics after training, e.g., the current learning rate,
-            gradient norm, etc.
-        """
-        raise NotImplementedError()
-
-    @torch.no_grad()
-    @abc.abstractmethod
-    def eval_batch(
-        self,
-        input_: dict[str, Any],
-        loss_fn: Callable[..., torch.Tensor],
-        loss_weight_fn: Callable[[dict[str, Any]], torch.Tensor],
-    ) -> torch.Tensor | None:
-        """Evaluate the model using the forward pass and loss function.
-
-        Note
-        ----
-        The loss_fn should process packed 1D inputs, instead of 2D inputs.
-
-        Parameters
-        ----------
-        input_ : dict[str, Any]
-            The input data for model forward pass and the loss function.
-            Redundant entries are allowed.
-        loss_fn : Callable[..., torch.Tensor]
-            The loss function. For actor (is_critic=False), it receives
-            (logprobs, entropy, input_data). For critic (is_critic=True),
-            it receives (values, input_data). Returns a scalar normalized loss.
-        loss_weight_fn : Callable[[dict[str, Any]], torch.Tensor]
-            A function used to calculate the weight of each micro-batch. Since
-            loss_fn normalizes the loss for a micro-batch, we need a corresponding
-            weight for each micro-batch to normalize the loss globally. The weight
-            is usually the number of response tokens in the batch.
-
-        Returns
-        -------
-        torch.Tensor or None
-            A scalar loss or None. The evaluation statistics should be aggregated
-            with `stats_tracker`.
-        """
-        raise NotImplementedError()
-
-    @torch.no_grad()
-    @abc.abstractmethod
-    def forward_batch(
-        self,
-        input_: dict[str, Any],
-        output_seqlens: list[int] | None = None,
-        aggregate_fn: Callable[[list[Any]], Any] = torch.cat,
-    ) -> torch.Tensor:
-        """Run the forward pass or inference on the model.
-
-        Note
-        ----
-        This operation is gradient-free.
-
-        Parameters
-        ----------
-        input_ : dict[str, Any]
-            The input data for model forward pass. Redundant entries are allowed.
-        output_seqlens : list[int], optional
-            The desired output sequence lengths. If None, assumes that the output
-            has the same lengths as inputs, by default None.
-        aggregate_fn : Callable[[list[Any]], Any], optional
-            A function to aggregate micro-batched outputs, by default torch.cat.
-
-        Returns
-        -------
-        Any
-            For actor (is_critic=False): logprobs tensor aggregated by `aggregate_fn`.
-            For critic (is_critic=True): values tensor aggregated by `aggregate_fn`.
-        """
-        raise NotImplementedError()
-
-    @torch.no_grad()
-    def forward(
-        self,
-        input_: dict[str, Any],
-        output_seqlens: list[int] | None = None,
-        aggregate_fn: Callable[[list[Any]], Any] = torch.cat,
-    ) -> torch.Tensor:
-        return self.forward_batch(input_, output_seqlens, aggregate_fn)
-
-    @abc.abstractmethod
-    def export_stats(self) -> dict[str, float]:
-        """Export the statistics recorded in this engine process.
-
-        Note
-        ----
-        Statistics will be all-reduced across the data parallel group
-        and broadcasted from the last pipeline parallel stage.
-
-        Returns
-        -------
-        dict[str, float]
-            The exported scalar statistics.
-        """
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def onload(self) -> None:
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def offload(self) -> None:
-        raise NotImplementedError()
+class TrainEngine(
+    TrainEngineDistMixin,
+    TrainEngineStateMixin,
+    TrainEngineComputeMixin,
+    abc.ABC,
+):
+    pass
 
 
 class InferenceEngine(abc.ABC):
