+++ areal/api/train_engine/dist.py
@@ -1,0 +1,105 @@
+from __future__ import annotations
+
+import abc
+
+import torch.distributed as dist
+
+from areal.api.alloc_mode import ParallelStrategy
+
+
+class TrainEngineDistMixin(abc.ABC):
+    @abc.abstractmethod
+    def create_process_group(self, parallel_strategy: ParallelStrategy | None = None):
+        """Initialize PyTorch distributed communication groups.
+
+        Parameters
+        ----------
+        parallel_strategy : ParallelStrategy, optional
+            The parallel strategy configuration for distributed training, by default None
+        """
+        raise NotImplementedError()
+
+    @property
+    @abc.abstractmethod
+    def data_parallel_group(self) -> dist.ProcessGroup:
+        """Get the data parallel communication group of this engine.
+
+        Returns
+        -------
+        dist.ProcessGroup
+            The data parallel communication group
+        """
+        raise NotImplementedError()
+
+    @property
+    @abc.abstractmethod
+    def data_parallel_rank(self) -> int:
+        """Get the rank of the current process in the data parallel group.
+
+        Returns
+        -------
+        int
+            The rank of the current process in the data parallel group
+        """
+        raise NotImplementedError()
+
+    @property
+    @abc.abstractmethod
+    def data_parallel_world_size(self) -> int:
+        """Get the world size of the data parallel group.
+
+        Returns
+        -------
+        int
+            The world size of the data parallel group
+        """
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def current_data_parallel_head(self) -> int:
+        """Get the current data parallel head rank.
+
+        Returns
+        -------
+        int
+            The rank of the current data parallel head
+        """
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def is_data_parallel_head(self) -> bool:
+        """Check if the current rank is the data parallel head of the current engine.
+
+        Returns
+        -------
+        bool
+            True if the current rank is the data parallel head, False otherwise
+        """
+        raise NotImplementedError()
+
+    @property
+    @abc.abstractmethod
+    def context_and_model_parallel_group(self) -> dist.ProcessGroup:
+        """Get the context and model parallel communication group of this engine.
+
+        Returns
+        -------
+        dist.ProcessGroup
+            The context and model parallel communication group
+        """
+        raise NotImplementedError()
+
+    @property
+    @abc.abstractmethod
+    def cpu_group(self) -> dist.ProcessGroup:
+        """Get the CPU communication group of this engine.
+
+        Returns
+        -------
+        dist.ProcessGroup
+            The CPU communication group
+        """
+        raise NotImplementedError()
+
+    def destroy(self):
+        """Destroy the engine and release GPU memory of models."""
